+++ app/(dashboard)/(routes)/video/page.tsx
@@ -41,6 +41,11 @@ const VideoPage = () => {
     catch(error:any){
       if (error?.response?.status === 403) {
         ProModal.onOpen();
+      }else if(error?.response?.status === 429){
+        toast.error('Oops! your openAI free credits might have used up or expired...')
+      }
+      else if(error?.response?.status === 500){
+        toast.error('Oops! Seems like my Open AI free credits expired!');
       } else {
         toast.error('Something Went Wrong');
       }